requirements.txt
requests>=2.22.0
python-telegram-bot
